Encapsulation
The salary attribute is private, which have restricted access.

The public setSalary() method takes a parameter (s) and assigns it to the salary attribute (salary = s).

The public getName(), getAge(), and getID() methods return the value of the private name, age, and id attributes.

Inside main(), we create an object of the Student class. Now we can use the setName(), setAge(), and setID() methods to set the value of the private attribute to MR XYZ(student name), 21(Student Age), and 20-101(Student Id). Then we call the getName(), getAge(), and getID() methods on the object to return the value.