Encapsulation
The salary attribute is private, which have restricted access.

The public setSalary() method takes a parameter (s) and assigns it to the salary attribute (salary = s).




Here, Box is class name


The public getHeight(),and getWidth() methods return the value of the private height and width attributes.

Inside main(), we create an object of the Box class. Now we can use the setHeight() and setWidth() methods to set the value of the private attribute to 7(height) and 4(width). Then we call the getHeight() and getWidth() methods on the object to return the value.